+++ docker-compose.yml
@@ -13,6 +13,7 @@ services:
       - mongo2
       - mongo3
       - rabbitmq
+      
   rabbitmq:
     image: rabbitmq:3-management-alpine
     container_name: rabbitmq
@@ -46,4 +47,5 @@ services:
       - 27019:27017
     restart: always
     entrypoint: [ "/usr/bin/mongod", "--bind_ip_all", "--replSet", "rs0" ]
+
