+++ src/responseProducer/parser/permission.parser.middleware.ts
@@ -14,12 +14,17 @@ export interface PermissionDoc {
   updated_at: string;
 }
 
+/**
+ * permissionIndexParser - permission msg parser for index queue
+ * @param data  - data object type (from mongo change)
+ * @returns FileResponse|undefiend - file formatted msg
+ */
 export function permissionIndexParser(data: DataObjectType): FileResponse | undefined {
   if (concludeMongoOperation(data.operation) !== OperationType.DELETE) {
     const permissionDoc: PermissionDoc = <PermissionDoc>data.fullDocument;
     const formattedData: FileResponse = new FileResponse({
       pushObjectReq: { event: OperationType.PERMISSIONS_CHANGE },
-      fileID: permissionDoc.fileID
+      fileId: permissionDoc.fileID
     });
 
     return formattedData;
@@ -28,11 +33,18 @@ export function permissionIndexParser(data: DataObjectType): FileResponse | unde
   return;
 }
 
+/**
+ * permissionHiParser - permission msg parser for hi queue
+ * @param data  - data object type (from mongo change)
+ * @returns DefaultResponse|undefiend - default formatted msg
+ */
 export function permissionHiParser(data: DataObjectType): DefaultResponse | undefined {
   const permissionDoc: PermissionDoc = <PermissionDoc>data.fullDocument;
   let operation: OperationType = concludeMongoOperation(data.operation);
+
+  // If the creator isn't as the dest user - it is a new share permission
   if (operation === OperationType.CREATE &&
-    permissionDoc.creator !== permissionDoc.userID) {
+      permissionDoc.creator !== permissionDoc.userID) {
     operation = OperationType.SHARE_CREATED;
 
     const formattedData: DefaultResponse = new DefaultResponse({
